tests/wpt/tests/css/css-animations/parsing/animation-trigger-exit-range-start-computed.html

<!DOCTYPE html>
<!-- TODO(crbug.com/390314945): Replace with spec link when the spec lands. -->
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/8942 ">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<div id="target" style="font-size:10px;"></div>
<script>
// https://drafts.csswg.org/css-animations-2/#animation-trigger-exit-range-start
test_computed_value("animation-trigger-exit-range-start", "initial", "auto");
test_computed_value("animation-trigger-exit-range-start", "auto");
test_computed_value("animation-trigger-exit-range-start", "normal");
test_computed_value("animation-trigger-exit-range-start", "cover 0%", "cover");
test_computed_value("animation-trigger-exit-range-start", "cover 100%");
test_computed_value("animation-trigger-exit-range-start", "COVER 0%", "cover");
test_computed_value("animation-trigger-exit-range-start", "COVER 100%", "cover 100%");
test_computed_value("animation-trigger-exit-range-start", "cover 120%");
test_computed_value("animation-trigger-exit-range-start", "cover 42%");
test_computed_value("animation-trigger-exit-range-start", "0", "0px");
test_computed_value("animation-trigger-exit-range-start", "120%");
test_computed_value("animation-trigger-exit-range-start", "120px");
test_computed_value("animation-trigger-exit-range-start", "cover -42%");
test_computed_value("animation-trigger-exit-range-start", "contain 42%");
test_computed_value("animation-trigger-exit-range-start", "exit 42%");
test_computed_value("animation-trigger-exit-range-start", "exit calc(41% + 1%)", "exit 42%");
test_computed_value("animation-trigger-exit-range-start", "exit 1%, cover 2%, contain 0%", "exit 1%, cover 2%, contain");
test_computed_value("animation-trigger-exit-range-start", "exit 1%, cover 2%, contain 100%");
test_computed_value("animation-trigger-exit-range-start", "exit-crossing 42%");
test_computed_value("animation-trigger-exit-range-start", "entry 42px");
test_computed_value("animation-trigger-exit-range-start", "entry-crossing 42px");
test_computed_value("animation-trigger-exit-range-start", "contain calc(10% + 10px)");
test_computed_value("animation-trigger-exit-range-start", "entry 1em", "entry 10px");
test_computed_value("animation-trigger-exit-range-start", "exit calc(1em + 10px)", "exit 20px");
test_computed_value("animation-trigger-exit-range-start", "cover");
test_computed_value("animation-trigger-exit-range-start", "contain");
test_computed_value("animation-trigger-exit-range-start", "entry");
test_computed_value("animation-trigger-exit-range-start", "exit");
</script>
